feat: made the hide command a function
